Move valid_conf to channels.py
